ref: move mod.rs code into layer.rs
